Implement automatic ESP discovery and sync
Replace the manual Scan/Sync flow with a background auto-discovery loop that scans continuously, authorizes devices (stub accepts all), and syncs readings automatically. Scan tab becomes a passive device list; Sync tab removed. Sensors tab now reads real data from SQLite via a devices table. Fix stale BLE characteristic cache across reconnects, merge API mismatches, and the shell routing crash.

+++ Protocol/EspControlEncoder.cs
@@ -16,7 +16,7 @@ public static class EspControlEncoder
 
     public static byte[] Cancel()
     {
-        return [EspProtocol.ControlCommand.Cancel, EspProtocol.Version];
+        return [(byte)EspProtocol.ControlCommand.Cancel, EspProtocol.Version];
     }
 
     public static byte[] BuildCommand(EspProtocol.ControlCommand command, uint bootSessionId, uint readingId)
